requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
PyYAML==6.0.1
python-dotenv==1.0.0
httpx==0.25.2
google-genai
aiofiles==23.2.1
supabase==2.3.4
